Rename files for improved import
Rename the elements.py to _base_elements.py as this should be a hidden module
with methods/classes for all elements. This should not be imported by the end
user.

Rename the beam.py to elements.py. This will be the module for all complete
elements. This will make importing easier, since they will all be in the same
place. A simple from femethods.elements import Beam, Bar, etc to import
multiple types of elements.

Update imports for exact methods.