Add load location validation; update solve method; invalidate on change
When a load a reaction are at the exact location, the global stiffness
matrix isn't calculated correctly. the new
Element.__validate_load_locations method will shift the loads towards
the inside of the beam by a small amount to allow the stiffness matrix
to be calculated properly.

Update the solve method to remesh automatically.

Invalidate results anytime the loads/reactions are changed.