Update plot method to allow customization of diagrams
The diagrams that the plot method can plot are the shear, moment, and
deflection diagrams. These can now be customized by passing name of the desired
diagrams to the `diagrams` parameter.
Custom `y-labels` can also be set by passing a tuple of `y-labels` to the
`diagram_labels` parameter.

The `bending_stress` parameter has been removed. This is no longer a valid
diagram.